Field value publishing shouldn't await subscribers
In case of published fields, the subscribers will only be interested in
the latest value. So, we don't need to wait for the subscribers to
receive the previous value in the buffer.

On the other hand, signals are themselves the data that the subscribers
are interested in, so it makes sense to wait for the subscribers to
receive every signal emission.